# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Szkola Reacta API",
    "description": "Szkola Reacta endpoints",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "api.szkolareacta.pl",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/events/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of event we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get event with given ID",
        "responses": {
          "200": {
            "description": "Event is found",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Event with given ID",
        "tags": [
          "events"
        ],
        "responses": {
          "200": {
            "description": "Event was deleted",
            "schema": {
              "$ref": "#/definitions/Empty"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update event with given ID",
        "tags": [
          "events"
        ],
        "parameters": [
          {
            "name": "post",
            "in": "body",
            "description": "Event with new values of properties",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event was updated",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/places/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of place we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get place with given ID",
        "responses": {
          "200": {
            "description": "Place is found",
            "schema": {
              "$ref": "#/definitions/Place"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Place with given ID",
        "tags": [
          "places"
        ],
        "responses": {
          "200": {
            "description": "Place was deleted",
            "schema": {
              "$ref": "#/definitions/Empty"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update place with given ID",
        "tags": [
          "places"
        ],
        "parameters": [
          {
            "name": "post",
            "in": "body",
            "description": "Place with new values of properties",
            "schema": {
              "$ref": "#/definitions/Place"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Place was updated",
            "schema": {
              "$ref": "#/definitions/Place"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Empty": {
      "properties": {

      }
    },
...
    "Event": {
      "required": [
        "title",
        "description",
        "price",
        "city",
        "category",
        "date"
      ],
      "properties": {
        "id": {
          "type": "string",
          "uniqueItems": true
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "cover_url": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "city": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "hour": {
          "type": "integer"
        }
      }
    },
...
    "Place": {
      "required": [
        "title",
        "description",
        "price",
        "city",
        "category",
        "date"
      ],
      "properties": {
        "id": {
          "type": "string",
          "uniqueItems": true
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "foto_url": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "city": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      }
    },
...
  }
}